Update Component hierarchy in activetrade
