+++ src/activetrade/layouts/ActiveTradeProgress.js
@@ -5,11 +5,8 @@ class ActiveTradeProgress extends Component {
   render() {
     // NOTE / TODO: the status here is mocked for now. It should probably
     // be fetched from our db for a particular transaction
-    const progress_map = [
-      { status: 'completed', label: '1', text: 'Escrow' },
-      { status: 'active', label: '2', text: 'Payment' },
-      { status: '', label: '3', text: 'Ether Released' }
-    ]
+    const progress_map = this.props.progress_map;
+
     return(
       <Progress map={progress_map} />
     )
